backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
SQLAlchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.2
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
python-multipart>=0.0.9
orjson>=3.10.0
tenacity>=9.0.0
